clear Chess and added Game classes
