+++ src/main/java/chess/chessjavafx/GameController.java
@@ -1,0 +1,14 @@
+package chess.chessjavafx;
+
+import chess.chessjavafx.JavaFX.GameScene;
+import chess.chessjavafx.Pieces.Piece;
+
+public class GameController {
+    private Checkerboard checkerboard;
+    private Piece.Team currentPlayer;
+    private Position selectedPiece;
+    private GameMoves moveData;
+    private GameScene gameScene;
+
+
+}
