+++ src/main/java/chess/chessjavafx/GameMoves.java
@@ -1,0 +1,4 @@
+package chess.chessjavafx;
+
+public class GameMoves {
+}
