+++ src/main/java/chess/chessjavafx/JavaFX/GameScene.java
@@ -1,0 +1,13 @@
+package chess.chessjavafx.JavaFX;
+
+import javafx.scene.Group;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.paint.Color;
+
+public class GameScene extends Scene {
+
+    public GameScene(Parent parent) {
+        super(new Group(), Color.LIME);
+    }
+}
